implement functionality related to spacebar. closes issue #4
the logic for finding the index of the next fillable square is updated as follows:
for arrow keys, we were using the following logic to select the next active square:
a. the index of the square should be higher than the currently active square
b. the square should be fillable
c. depending on the active direction, the square should be in the same row or column as the current active square
it did not matter whether the next square was empty or had previously been filled by the user

for spacebar, the additional conditions were:
a. if the currently active square is empty, the next active square has to be empty as well
b. if the currently active square is already filled, then pressing the spacebar will delete its contents and move to the next square whether its empty or filled